requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
Flask-Login
pymongo
python-dotenv
requests